Example connection string update
The main example connection string (where you can click "Copy" on the public docs page) is confusing, as evidenced by recent casework.

A customer came in with a connection string like this (none of this is real customer data):
jdbc:sqlserver://;serverName=10.0.0.1;databaseName=testDB;AuthentificationMethod=ntlm;integratedSecurity=true

There should be no semi-colon at the beginning, before serverName. The servername should literally be just the servername, not serverName=server_name.

AuthentificationMethod is something we'll address with the customer of course, it's not relevant to this article update.