Update listeners-client-connectivity-application-failover.md
I propose to change "setspn -A" on line 182 to "setspn -S". I realize the functionality as been updated, but we still have people out there with older Operating Systems where it is not yet updated.

I love that the functionality is updated, and eventually this shouldn't matter, but I would prefer we put the 100% consistent "setspn -S" in place to avoid issues.